+++ ZZZ-Optimizer/app/page.tsx
@@ -2,8 +2,8 @@
 
 import { TextArt } from "@/components/textArt";
 import { siteConfig } from "@/config/site";
-import { UploadButton } from "@/components/uploadButton";
-import { ScanDataTerminalOutput } from "@/components/scanDataTerminalOutput";
+import { UploadButton } from "@/components/landing-page/uploadButton";
+import { ScanDataTerminalOutput } from "@/components/landing-page/scanDataTerminalOutput";
 import { useScanStore } from "@/atomsAndStores/useScanStore";
 
 export default function Home() {
