+++ ZZZ-Optimizer/lib/artifactStats.ts
@@ -1,0 +1,48 @@
+const valid_set_names = [
+  "Swing Jazz",
+  "Chaotic Metal",
+  "Hormone Punk",
+  "Fanged Metal",
+  "Shockstar Disco",
+  "Thunder Metal",
+  "Woodpecker Electro",
+  "Soul Rock",
+  "Puffer Electro",
+  "Inferno Metal",
+  "Freedom Blues",
+  "Polar Metal",
+  "Astral Voice",
+  "Branch & Blade Song",
+  "Chaos Jazz",
+  "Proto Punk",
+];
+
+const valid_base_stats = [
+  "HP",
+  "ATK",
+  "DEF",
+  "CRIT Rate",
+  "CRIT DMG",
+  "Anomaly Proficiency",
+  "PEN Ratio",
+  "Physical DMG Bonus",
+  "Fire DMG Bonus",
+  "Ice DMG Bonus",
+  "Electric DMG Bonus",
+  "Ether DMG Bonus",
+  "Anomaly Mastery",
+  "Impact",
+  "Energy Regen",
+];
+
+const valid_random_stats = [
+  "HP",
+  "ATK",
+  "DEF",
+  "CRIT Rate",
+  "CRIT DMG",
+  "Anomaly Proficiency",
+  "PEN",
+];
+
+export { valid_set_names, valid_base_stats, valid_random_stats };
